sound yayy

+++ main.py
@@ -1,46 +1,45 @@
+import sys
+
 from screen import run_all
 from screen import play_sound
 import multiprocessing
-import time
-from multiprocessing.pool import ThreadPool
+from multiprocessing import Manager
 import config
 
 config.init()
 mylist = [ False]
 
-def manager():
+def soundmanager(num):
     p3 = multiprocessing.Process(target=play_sound)
     p3.start()
+    count = 1
     # p3.terminate()
     while(True):
-
-        print(mylist[0])
-        time.sleep(1)
-        if mylist[0] == True:
+        if num.value == 1:
             p3.terminate()
-
-def music():
-    p2 = multiprocessing.Process(target=play_sound)
-    p2.start()
-    while True:
-        if config._finish:
-            print("frick uuuu")
-            break
-        p2.join()
-    print("terminated")
-    p2.terminate()
+            count -= 1
+            num.value = 2
+        if num.value == 0 and count == 0:
+            p3 = multiprocessing.Process(target=play_sound)
+            p3.start()
+            num.value = 2
+            count += 1
 
 
-if __name__ == '__main__':
-    # pool = ThreadPool(processes=2)
-    # pool.apply_async(run_all)
-    # pool.apply_async(play_sound)
-    # FINISH = True
-    # pool.terminate()
-    # pool.join()
-    p1 = multiprocessing.Process(target=run_all, args=(mylist, ))
-    p2 = multiprocessing.Process(target=manager)
+def processManager():
+    manager = Manager()
+    num = manager.Value('i', 2)
+    finish = manager.Value('i',0)
+
+    p1 = multiprocessing.Process(target=run_all, args=(num,finish,))
+    p2 = multiprocessing.Process(target=soundmanager, args=(num,))
     p1.start()
     p2.start()
-    p1.join()
-    p2.join()
+    while(True):
+        if finish.value == 1:
+            p1.terminate()
+            p2.terminate()
+            sys.exit()
+
+if __name__ == '__main__':
+    processManager()
